requirements.txt
Django==2.2.6
feedparser==5.2.1
psycopg2==2.8.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
